Labfiles/04-azure-speech/Python/voice-mail/requirements.txt
python-dotenv
playsound3
azure-identity
azure-cognitiveservices-speech==1.48.2